Time: 10 ms (95.34%), Space: 88.1 MB (73.33%) - LeetHub

+++ 2226-maximum-candies-allocated-to-k-children/2226-maximum-candies-allocated-to-k-children.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    bool isPossible(vector<int> &candies, int n,long long children){
+    bool isPossible(vector<int> &candies, long long n,long long children){
         for(auto &i : candies){
             children -= i/n;
             if(children <= 0)   return true;
